IDEMPIERE-6989 Distributed Cache: Redis - reorganise into sub-packages by concern
Moves the 7 internal classes into sub-packages so the bundle's
OSGi-public surface (the 3 DS impls + Activator) stays at the top-level
package while implementation details live in clearly-named internals:

  config/  - RedisConfig
  health/  - RedisHealth
  cache/   - CaffeineLayeredMap
  message/ - TopicImpl
  cluster/ - ClusterMember, RpcRequest, RpcResponse

No behavioural change - pure file moves + import statement updates.
The 3 @Component classes (CacheServiceImpl, MessageServiceImpl,
ClusterServiceImpl) and the Activator stay at the top-level package
so OSGI-INF DS XML class names remain stable.
